+++ src/tp/logic/BruteForce.java
@@ -123,7 +123,7 @@ public class BruteForce {
 		return (double) totalRating / combination.size();
 	}
 
-	public int getCombinationCount() {
+	public double getCombinationCount() {
 		return combinationCount;
 	}
